requirements.txt
flask
python-dotenv
mysql-connector-python
bs4
pdfplumber
elasticsearch
selenium
outscraper
python-dateutil
APScheduler
fake_useragent
webdriver-manager